% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooarchGUI.r
\docType{data}
\name{ClarksonLimbLoc}
\alias{ClarksonLimbLoc}
\title{Clarkson data for chi squared}
\description{
Clarkson data for chi squared
}
\references{
Otarola-Castillo, Erik 2010 Differences between NISP and MNE in Cutmark Analysis of Highly Fragmented Faunal Assemblages. Journal of Archaeological Science 37:1-12.
}
\author{
Erik Otarola-Castillo
}
\keyword{datasets}
