% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooarchGUI.r
\docType{data}
\name{clusterPTS}
\alias{clusterPTS}
\title{Clustered Spatial Points Example for spatial statistics and autocorrelation example}
\description{
Clustered Spatial Points Example for spatial statistics and autocorrelation example
}
\author{
Erik Otarola-Castillo
}
\keyword{datasets}
