% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooarchGUI.r
\docType{data}
\name{LRBdata}
\alias{LRBdata}
\title{Forager databse from Binford 2001, Constructing Frames of Reference}
\description{
Forager databse from Binford 2001, Constructing Frames of Reference
}
\references{
Binford, Lewis R. 2001. Constructing Frames of Reference: An Analytical Method for Archaeological Theory Building Using Ethnographic and Environmental Data Sets. Berkeley: University of California Press.

Available from http://ajohnson.sites.truman.edu/data-and-program/
}
\author{
Binford, Lewis R.
}
\keyword{datasets}
