% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooarchGUI.r
\docType{data}
\name{marjRab.fuse}
\alias{marjRab.fuse}
\title{Fusion Survival data from Marj Rabba}
\description{
Fusion Survival data from Marj Rabba
}
\references{
Price, M.D., Buckley, M., Rowan, Y.M., Kersel, M., 2013. Animal Management Strategies during the Chalcolithic in the Lower Galilee: New Data from Marj Rabba, Paleorient 39, 183 - 200.
}
\author{
M.D. Price, M. Buckley, Y.M. Rowan, and M. Kersel
}
\keyword{datasets}
