% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooarchGUI.r
\docType{data}
\name{marjRab}
\alias{marjRab}
\title{Survival data from Marj Rabba, using Payne's (1973) age classes}
\description{
Survival data from Marj Rabba, using Payne's (1973) age classes
}
\references{
Payne, S. 1973  Kill - off Patterns in Sheep and Goats: The Mandibles from Asvan Kale. Anatolian Studies 23:281 - 303.

Price, M.D., Buckley, M., Rowan, Y.M., Kersel, M., 2013. Animal Management Strategies during the Chalcolithic in the Lower Galilee: New Data from Marj Rabba, Paleorient 39, 183 - 200.
}
\author{
M.D. Price, M. Buckley, Y.M. Rowan, and M. Kersel
}
\keyword{datasets}
