% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooarchGUI.r
\docType{data}
\name{winslow.fuse}
\alias{winslow.fuse}
\title{Fusion Survival data for cattle remains from the 
Winslow site, a colonial period farm near Boston, MA.}
\description{
Fusion Survival data for cattle remains from the 
Winslow site, a colonial period farm near Boston, MA.
}
\references{
Landon, David B. 1993	Feeding Colonial Boston: A Zooarchaeological Study. Historical Archaeology 30:i-vii, 1-153
}
\author{
D. Landon
}
\keyword{datasets}
