% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooarchGUI.r
\docType{package}
\name{zooaRchGUI-package}
\alias{zooaRchGUI-package}
\alias{zooaRchGUI}
\title{Analytical Tools for Zooarchaeological Data}
\description{
Functions in this package allow users to interactively read, manage, analyze and visualize
zooarchaeological data - the faunal remains recovered from archaeological sites.
}
\author{
Erik Otarola-Castillo, Jesse Wolfhagen, John Brian Rapes and Max D. Price
}
